+++ packages/typespec-client-generator-core/test/types.test.ts
@@ -165,7 +165,7 @@ describe("typespec-client-generator-core: types", () => {
         }
       `
       );
-      const models = getAllModels(runnerWithCore.context);
+      const models = runnerWithCore.context.experimental_sdkPackage.models;
       for (const property of (models[0] as SdkModelType).properties) {
         strictEqual(
           property.type.kind,
@@ -250,11 +250,11 @@ describe("typespec-client-generator-core: types", () => {
         ): void;
       `
       );
-      const [models, diagnostics] = getAllModelsWithDiagnostics(runner.context);
+      const diagnostics = runner.context.experimental_sdkPackage.diagnostics;
       expectDiagnostics(diagnostics, []);
-      const m = models.find((x) => x.name === "TestModel")!;
-      const e1 = models.find((x) => x.name === "TestEnum")!;
-      const e2 = models.find((x) => x.name === "testScalar")!;
+      const m = runner.context.experimental_sdkPackage.models.find((x) => x.name === "TestModel")!;
+      const e1 = runner.context.experimental_sdkPackage.enums.find((x) => x.name === "TestEnum")!;
+      const e2 = runner.context.experimental_sdkPackage.enums.find((x) => x.name === "testScalar")!;
       strictEqual(e1.kind, "enum");
       strictEqual(e1.valueType.kind, "string");
       strictEqual(e2.kind, "enum");
@@ -541,7 +541,7 @@ describe("typespec-client-generator-core: types", () => {
       }
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 4);
       const modelWithNamedUnionProperty = models.find(
         (x) => x.kind === "model" && x.name === "ModelWithNamedUnionProperty"
@@ -602,7 +602,7 @@ describe("typespec-client-generator-core: types", () => {
       }
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const model = models.find((x) => x.kind === "model" && x.name === "Test")! as SdkModelType;
       const sdkType = model.properties[0].type;
@@ -632,7 +632,7 @@ describe("typespec-client-generator-core: types", () => {
       }
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 3);
       const model = models.find((x) => x.kind === "model" && x.name === "Test")! as SdkModelType;
       const nullableModel = models.find(
@@ -680,9 +680,9 @@ describe("typespec-client-generator-core: types", () => {
       }
       `);
 
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 2);
-      const sdkType = models.find((x) => x.kind === "enum")! as SdkEnumType;
+      strictEqual(runner.context.experimental_sdkPackage.models.length, 1);
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const sdkType = runner.context.experimental_sdkPackage.enums[0];
       strictEqual(sdkType.isFixed, false);
       strictEqual(sdkType.name, "DaysOfWeekExtensibleEnum");
       strictEqual(sdkType.valueType.kind, "string");
@@ -730,9 +730,9 @@ describe("typespec-client-generator-core: types", () => {
       }
       `);
 
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 2);
-      const sdkType = models.find((x) => x.kind === "enum")! as SdkEnumType;
+      strictEqual(runner.context.experimental_sdkPackage.models.length, 1);
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const sdkType = runner.context.experimental_sdkPackage.enums[0];
       strictEqual(sdkType.isFixed, false);
       strictEqual(sdkType.name, "Integers");
       strictEqual(sdkType.valueType.kind, "int32");
@@ -893,9 +893,9 @@ describe("typespec-client-generator-core: types", () => {
           }
         }
       `);
-      const models = getAllModels(runnerWithCore.context);
-      strictEqual(models.length, 2);
-      const sdkType = models.find((x) => x.kind === "enum")! as SdkEnumType;
+      strictEqual(runnerWithCore.context.experimental_sdkPackage.models.length, 1);
+      strictEqual(runnerWithCore.context.experimental_sdkPackage.enums.length, 1);
+      const sdkType = runnerWithCore.context.experimental_sdkPackage.enums[0];
       strictEqual(sdkType.isFixed, true);
       strictEqual(sdkType.name, "DaysOfWeekFixedEnum");
       strictEqual(sdkType.valueType.kind, "string");
@@ -938,9 +938,7 @@ describe("typespec-client-generator-core: types", () => {
         ): void;
       `);
 
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 1);
-      strictEqual(models[0].access, "internal");
+      strictEqual(runner.context.experimental_sdkPackage.enums[0].access, "internal");
     });
     it("crossLanguageDefinitionId", async () => {
       await runner.compile(`
@@ -963,9 +961,8 @@ describe("typespec-client-generator-core: types", () => {
           }
         }
       `);
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 2);
-      const integersEnum = models.find((x) => x.kind === "enum")! as SdkEnumType;
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const integersEnum = runner.context.experimental_sdkPackage.enums[0];
       strictEqual(integersEnum.crossLanguageDefinitionId, "MyService.Integers");
     });
 
@@ -982,9 +979,7 @@ describe("typespec-client-generator-core: types", () => {
         ): void;
       `);
 
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 1);
-      strictEqual(models[0].deprecation, "no longer support");
+      strictEqual(runner.context.experimental_sdkPackage.enums[0].deprecation, "no longer support");
     });
 
     it("orphan enum", async () => {
@@ -1008,9 +1003,11 @@ describe("typespec-client-generator-core: types", () => {
         }
       `);
 
-      const models = getAllModels(runner.context);
-      strictEqual(models[0].name, "Enum1");
-      strictEqual(models[0].usage, UsageFlags.Input | UsageFlags.Output);
+      strictEqual(runner.context.experimental_sdkPackage.enums[0].name, "Enum1");
+      strictEqual(
+        runner.context.experimental_sdkPackage.enums[0].usage,
+        UsageFlags.Input | UsageFlags.Output
+      );
     });
 
     it("projected name", async () => {
@@ -1156,7 +1153,7 @@ describe("typespec-client-generator-core: types", () => {
         }
       `);
 
-      const sdkModel = getAllModels(runner.context)[0] as SdkModelType;
+      const sdkModel = runner.context.experimental_sdkPackage.models[0] as SdkModelType;
       strictEqual(sdkModel.name, "JavaTest");
 
       // Java projected name test
@@ -1235,7 +1232,7 @@ describe("typespec-client-generator-core: types", () => {
           removedProp: string;
         }
       `);
-      const sdkModel = getAllModels(runner.context).find(
+      const sdkModel = runner.context.experimental_sdkPackage.models.find(
         (x) => x.kind === "model"
       )! as SdkModelType;
       strictEqual(sdkModel.kind, "model");
@@ -1310,7 +1307,7 @@ describe("typespec-client-generator-core: types", () => {
           op test(@body input: InputModel): OutputModel;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["InputModel", "OutputModel"].sort());
@@ -1327,7 +1324,7 @@ describe("typespec-client-generator-core: types", () => {
           op test(@body input: Record<InnerModel>): void;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["InnerModel"].sort());
@@ -1344,7 +1341,7 @@ describe("typespec-client-generator-core: types", () => {
           op test(@body input: InnerModel[]): void;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["InnerModel"].sort());
@@ -1365,7 +1362,7 @@ describe("typespec-client-generator-core: types", () => {
           op test(@body input: InputModel): void;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["InputModel", "InnerModel"].sort());
@@ -1386,7 +1383,7 @@ describe("typespec-client-generator-core: types", () => {
           op test(@body input: InputModel): void;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["InputModel", "BaseModel"].sort());
@@ -1404,7 +1401,7 @@ describe("typespec-client-generator-core: types", () => {
 
       op test(@body input: DerivedModel): void;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["InputModel", "DerivedModel"].sort());
@@ -1418,7 +1415,7 @@ describe("typespec-client-generator-core: types", () => {
         prop: RecursiveModel
       }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const recursiveModel = models[0];
       strictEqual(recursiveModel.name, "RecursiveModel");
@@ -1462,7 +1459,7 @@ describe("typespec-client-generator-core: types", () => {
       @get
       op getModel(): Fish;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 5);
       const fish = models.find((x) => x.name === "Fish")! as SdkModelType;
       const kindProperty = fish.properties.find(
@@ -1490,7 +1487,7 @@ describe("typespec-client-generator-core: types", () => {
       @get
       op getModel(): Fish;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const fish = models.find((x) => x.name === "Fish")! as SdkModelType;
       const kindProperty = fish.properties.find(
@@ -1522,8 +1519,8 @@ describe("typespec-client-generator-core: types", () => {
       @get
       op getExtensibleModel(): Dog;
       `);
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 3);
+      const models = runner.context.experimental_sdkPackage.models;
+      strictEqual(models.length, 2);
 
       const golden = models.find((x) => x.name === "Golden")! as SdkModelType;
 
@@ -1533,7 +1530,8 @@ describe("typespec-client-generator-core: types", () => {
       strictEqual((kind.type as SdkEnumValueType).value, "golden");
 
       const dog = models.find((x) => x.name === "Dog")! as SdkModelType;
-      const dogKind = models.find((x) => x.name === "DogKind")! as SdkEnumType;
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const dogKind = runner.context.experimental_sdkPackage.enums[0];
 
       const dogKindProperty = dog.properties.find(
         (x) => (x as SdkBodyModelPropertyType).serializedName === "kind"
@@ -1555,9 +1553,8 @@ describe("typespec-client-generator-core: types", () => {
       @put
       op putPet(@body petKind: PetKind): void;
       `);
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 1);
-      const petKind = models[0] as SdkEnumType;
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const petKind = runner.context.experimental_sdkPackage.enums[0];
       strictEqual(petKind.name, "PetKind");
       strictEqual(petKind.isFixed, false);
       strictEqual(petKind.valueType.kind, "string");
@@ -1616,10 +1613,12 @@ describe("typespec-client-generator-core: types", () => {
       @put
       op putPet(@body pet: Pet): void;
       `);
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 2);
-      const pet = models.find((x) => x.name === "Pet")! as SdkModelType;
-      const kind = models.find((x) => x.name === "")!;
+      const models = runner.context.experimental_sdkPackage.models;
+      strictEqual(models.length, 1);
+      const pet = models.find((x) => x.name === "Pet")!;
+
+      const enums = runner.context.experimental_sdkPackage.enums;
+      const kind = enums.find((x) => x.name === "")!;
       strictEqual(kind.generatedName, "PetKind");
       const kindProperty = pet.properties.find(
         (x) => (x.nameInClient = "kind")
@@ -1646,8 +1645,8 @@ describe("typespec-client-generator-core: types", () => {
       @get
       op getExtensibleModel(): Dog;
       `);
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 3);
+      const models = runner.context.experimental_sdkPackage.models;
+      strictEqual(models.length, 2);
 
       const golden = models.find((x) => x.name === "Golden")! as SdkModelType;
 
@@ -1657,7 +1656,8 @@ describe("typespec-client-generator-core: types", () => {
       strictEqual((kind.type as SdkEnumValueType).value, "golden");
 
       const dog = models.find((x) => x.name === "Dog")! as SdkModelType;
-      const dogKind = models.find((x) => x.name === "DogKind")! as SdkEnumType;
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const dogKind = runner.context.experimental_sdkPackage.enums[0];
 
       const dogKindProperty = dog.properties.find(
         (x) => (x as SdkBodyModelPropertyType).serializedName === "kind"
@@ -1696,7 +1696,7 @@ describe("typespec-client-generator-core: types", () => {
       @get
       op getModel(): Fish;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 5);
       const shark = models.find((x) => x.name === "Shark")! as SdkModelType;
       strictEqual(shark.properties.length, 2);
@@ -1733,20 +1733,24 @@ describe("typespec-client-generator-core: types", () => {
       @get
       op getModel(): Fish;
       `);
-      const models = getAllModels(runner.context);
-      strictEqual(models.length, 4);
-      const fish = models.find((x) => x.name === "Fish")! as SdkModelType;
+      const models = runner.context.experimental_sdkPackage.models;
+      strictEqual(models.length, 3);
+      const fish = models.find((x) => x.name === "Fish")!;
       let kindTypeProperty = fish.properties.find((x) => x.nameInClient === "kind")!;
       strictEqual(kindTypeProperty.type.kind, "enum");
-      const shark = models.find((x) => x.name === "Shark")! as SdkModelType;
+      const shark = models.find((x) => x.name === "Shark")!;
       strictEqual(shark.discriminatorValue, "shark");
       kindTypeProperty = shark.properties.find((x) => x.nameInClient === "kind")!;
       strictEqual(kindTypeProperty.type.kind, "enumvalue");
-      const salmon = models.find((x) => x.name === "Salmon")! as SdkModelType;
+      const salmon = models.find((x) => x.name === "Salmon")!;
       kindTypeProperty = salmon.properties.find((x) => x.nameInClient === "kind")!;
       strictEqual(kindTypeProperty.type.kind, "enumvalue");
       strictEqual(salmon.discriminatorValue, "salmon");
-      const kindType = models.find((x) => x.name === "KindType")! as SdkEnumType;
+
+      strictEqual(runner.context.experimental_sdkPackage.enums.length, 1);
+      const kindType = runner.context.experimental_sdkPackage.enums.find(
+        (x) => x.name === "KindType"
+      )!;
       strictEqual(kindType.isFixed, false);
     });
 
@@ -1777,7 +1781,7 @@ describe("typespec-client-generator-core: types", () => {
           op createOrUpdate is ResourceCreateOrUpdate<User>;
         }
       `);
-      const models = getAllModels(runnerWithCore.context);
+      const models = runnerWithCore.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].name, "User");
     });
@@ -1810,7 +1814,7 @@ describe("typespec-client-generator-core: types", () => {
           op createOrUpdate is ResourceCreateOrUpdate<User>;
         }
       `);
-      const models = getAllModels(runnerWithCore.context);
+      const models = runnerWithCore.context.experimental_sdkPackage.models;
       strictEqual(models.length, 4);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(modelNames, ["Error", "ErrorResponse", "InnerError", "User"].sort());
@@ -1844,7 +1848,7 @@ describe("typespec-client-generator-core: types", () => {
           op createOrUpdateUser is LongRunningResourceCreateOrUpdate<User>;
         }
       `);
-      const models = getAllModels(runnerWithCore.context);
+      const models = runnerWithCore.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].name, "User");
     });
@@ -1877,27 +1881,22 @@ describe("typespec-client-generator-core: types", () => {
           op createOrUpdateUser is LongRunningResourceCreateOrUpdate<User>;
         }
       `);
-      const models = getAllModels(runnerWithCore.context);
-      strictEqual(models.length, 6);
+      const models = runnerWithCore.context.experimental_sdkPackage.models;
+      strictEqual(models.length, 5);
       const modelNames = models.map((model) => model.name).sort();
       deepStrictEqual(
         modelNames,
-        [
-          "Error",
-          "ErrorResponse",
-          "InnerError",
-          "User",
-          "ResourceOperationStatus",
-          "OperationState",
-        ].sort()
+        ["Error", "ErrorResponse", "InnerError", "User", "ResourceOperationStatus"].sort()
       );
+      strictEqual(runnerWithCore.context.experimental_sdkPackage.enums.length, 1);
+      strictEqual(runnerWithCore.context.experimental_sdkPackage.enums[0].name, "OperationState");
     });
     it("no models filter core", async () => {
       await runner.compile(`
         @service({})
         @test namespace MyService { }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 0);
     });
     it("no models don't filter core", async () => {
@@ -1905,7 +1904,7 @@ describe("typespec-client-generator-core: types", () => {
         @service({})
         @test namespace MyService { }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 0);
     });
     it("input usage", async () => {
@@ -1915,12 +1914,11 @@ describe("typespec-client-generator-core: types", () => {
         }
         op operation(@body input: InputModel): void;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].usage, UsageFlags.Input);
-
-      strictEqual(getAllModels(runner.context, { input: true, output: false }).length, 1);
-      strictEqual(getAllModels(runner.context, { input: false }).length, 0);
+      strictEqual(models.filter((x) => x.usage === UsageFlags.Input).length, 1);
+      strictEqual(models.filter((x) => x.usage === UsageFlags.Output).length, 0);
     });
 
     it("output usage", async () => {
@@ -1930,12 +1928,12 @@ describe("typespec-client-generator-core: types", () => {
         }
         op operation(): OutputModel;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].usage, UsageFlags.Output);
 
-      strictEqual(getAllModels(runner.context, { input: false, output: true }).length, 1);
-      strictEqual(getAllModels(runner.context, { output: false }).length, 0);
+      strictEqual(models.filter((x) => x.usage === UsageFlags.Output).length, 1);
+      strictEqual(models.filter((x) => x.usage === UsageFlags.Input).length, 0);
     });
 
     it("roundtrip usage", async () => {
@@ -1945,13 +1943,46 @@ describe("typespec-client-generator-core: types", () => {
         }
         op operation(@body input: RoundtripModel): RoundtripModel;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].usage, UsageFlags.Input | UsageFlags.Output);
 
-      strictEqual(getAllModels(runner.context, { input: false, output: true }).length, 1);
-      strictEqual(getAllModels(runner.context, { input: true, output: false }).length, 1);
-      strictEqual(getAllModels(runner.context, { input: false, output: false }).length, 0);
+      strictEqual(models.filter((x) => (x.usage & UsageFlags.Output) > 0).length, 1);
+      strictEqual(models.filter((x) => (x.usage & UsageFlags.Input) > 0).length, 1);
+      strictEqual(models.filter((x) => x.usage === UsageFlags.None).length, 0);
+    });
+
+    it("usage propagation", async () => {
+      await runner.compileWithBuiltInService(`
+        @discriminator("kind")
+        model Fish {
+          age: int32;
+        }
+
+        @discriminator("sharktype")
+        model Shark extends Fish {
+          kind: "shark";
+        }
+
+        model Salmon extends Fish {
+          kind: "salmon";
+          friends?: Fish[];
+          hate?: Record<Fish>;
+          partner?: Fish;
+        }
+
+        model SawShark extends Shark {
+          sharktype: "saw";
+        }
+
+        model GoblinShark extends Shark {
+          sharktype: "goblin";
+        }
+        op operation(@body input: Shark): Shark;
+      `);
+      const models = runner.context.experimental_sdkPackage.models;
+      strictEqual(models.length, 5);
+      strictEqual(models[0].usage, UsageFlags.Input | UsageFlags.Output);
     });
 
     it("unnamed model", async () => {
@@ -1964,7 +1995,7 @@ describe("typespec-client-generator-core: types", () => {
           @body body: Test
         ): void;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 3);
       const propreties: string[] = [];
       models.forEach((model) => {
@@ -1986,7 +2017,7 @@ describe("typespec-client-generator-core: types", () => {
         ): void;
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].access, "internal");
     });
@@ -2043,7 +2074,7 @@ describe("typespec-client-generator-core: types", () => {
           @body body: Test6
         ): void;
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 6);
 
       const Test1 = models.find((x) => x.name === "Test1")!;
@@ -2082,7 +2113,7 @@ describe("typespec-client-generator-core: types", () => {
           prop: string;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 3);
       const AdditionalPropertiesModel = models.find(
         (x) => x.name === "AdditionalPropertiesModel"
@@ -2145,7 +2176,7 @@ describe("typespec-client-generator-core: types", () => {
           model OutputModel {}
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const inputModel = models.find((x) => x.name === "InputModel")!;
       strictEqual(inputModel.crossLanguageDefinitionId, "MyService.InputModel");
@@ -2160,7 +2191,6 @@ describe("typespec-client-generator-core: types", () => {
         model Catalog is TrackedResource<CatalogProperties> {
           @pattern("^[A-Za-z0-9_-]{1,50}$")
           @key("catalogName")
-          @path
           @segment("catalogs")
           name: string;
         }
@@ -2179,7 +2209,6 @@ describe("typespec-client-generator-core: types", () => {
         @access(Access.public)
         model Deployment is TrackedResource<DeploymentProperties> {
           @key("deploymentName")
-          @path
           @segment("deployments")
           name: string;
         }
@@ -2191,12 +2220,12 @@ describe("typespec-client-generator-core: types", () => {
           deploymentDateUtc?: utcDateTime;
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 4);
       const catalog = models.find((x) => x.name === "Catalog")! as SdkModelType;
       const deployment = models.find((x) => x.name === "Deployment")! as SdkModelType;
-      strictEqual(catalog.properties.length, 1);
-      strictEqual(deployment.properties.length, 1);
+      strictEqual(catalog.properties.length, 2);
+      strictEqual(deployment.properties.length, 2);
     });
     it("model with deprecated annotation", async () => {
       await runner.compileAndDiagnose(`
@@ -2210,7 +2239,7 @@ describe("typespec-client-generator-core: types", () => {
         ): void;
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].deprecation, "no longer support");
     });
@@ -2228,7 +2257,7 @@ describe("typespec-client-generator-core: types", () => {
         }
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       strictEqual(models[0].name, "Model1");
       strictEqual(models[0].usage, UsageFlags.Input | UsageFlags.Output);
@@ -2260,7 +2289,7 @@ describe("typespec-client-generator-core: types", () => {
           }
         }
       `);
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
     });
     it("error model", async () => {
@@ -2292,9 +2321,22 @@ describe("typespec-client-generator-core: types", () => {
           code: string
         };
 
-        model FourHundredError extends ApiError {};
-        model FourZeroFourError extends FourHundredError {};
-        model FiveHundredError extends ApiError {};
+        model FourHundredError extends ApiError {
+          @statusCode
+          @minValue(400)
+          @maxValue(499)
+          statusCode: int32;
+        };
+        model FourZeroFourError extends FourHundredError {
+          @statusCode
+          statusCode: 404;
+        };
+        model FiveHundredError extends ApiError {
+          @statusCode
+          @minValue(500)
+          @maxValue(599)
+          statusCode: int32;
+        };
 
         op test(): ValidResponse | FourZeroFourError | FiveHundredError;
       `);
@@ -2325,7 +2367,7 @@ describe("typespec-client-generator-core: types", () => {
       op basic(@header contentType: "multipart/form-data", @body body: MultiPartRequest): NoContentResponse;
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const model = models[0] as SdkModelType;
       strictEqual(model.kind, "model");
@@ -2372,7 +2414,7 @@ describe("typespec-client-generator-core: types", () => {
         @post op normalOperation(...B): void;
         `
       );
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
       const modelA = models.find((x) => x.name === "A")!;
       strictEqual(modelA.kind, "model");
@@ -2409,7 +2451,7 @@ describe("typespec-client-generator-core: types", () => {
         @put op multipartOne(@header contentType: "multipart/form-data", @body body: AddressFirstAppearance): void;
         `
       );
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 3);
     });
 
@@ -2423,7 +2465,7 @@ describe("typespec-client-generator-core: types", () => {
         @put op multipartOp(@header contentType: "multipart/form-data", @body body: PictureWrapper): void;
         `
       );
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const model = models[0] as SdkModelType;
       strictEqual(model.properties.length, 1);
@@ -2444,7 +2486,7 @@ describe("typespec-client-generator-core: types", () => {
         @put op multipartOp(@header contentType: "multipart/form-data", @body body: EncodedBytesMFD): void;
         `
       );
-      expectDiagnostics(getAllModelsWithDiagnostics(runner.context)[1], {
+      expectDiagnostics(runner.context.experimental_sdkPackage.diagnostics, {
         code: "@azure-tools/typespec-client-generator-core/encoding-multipart-bytes",
       });
     });
@@ -2464,7 +2506,7 @@ describe("typespec-client-generator-core: types", () => {
         @post op normalOp(): void | ErrorResponse;
         `
       );
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 2);
 
       const pictureWrapper = models.find((x) => x.name === "PictureWrapper")!;
@@ -2489,7 +2531,7 @@ describe("typespec-client-generator-core: types", () => {
         }
       `);
 
-      const models = getAllModels(runner.context);
+      const models = runner.context.experimental_sdkPackage.models;
       strictEqual(models.length, 1);
       const scopes = (models[0] as SdkModelType).properties.find(
         (x) => x.nameInClient === "scopes"
@@ -2509,7 +2551,9 @@ describe("typespec-client-generator-core: types", () => {
 });
 
 function getSdkBodyModelPropertyTypeHelper(runner: SdkTestRunner): SdkBodyModelPropertyType {
-  const sdkModel = getAllModels(runner.context).find((x) => x.kind === "model")! as SdkModelType;
+  const sdkModel = runner.context.experimental_sdkPackage.models.find(
+    (x) => x.kind === "model"
+  )! as SdkModelType;
   strictEqual(sdkModel.kind, "model");
   const property = sdkModel.properties[0];
   strictEqual(property.kind, "property");